Added annotations as attributes to CityJSON
