+++ 3DAmsterdam/Assets/AnnotationUI.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using ConvertCoordinates;
 using SimpleJSON;
 using UnityEngine;
 using UnityEngine.UI;
@@ -8,6 +9,7 @@ using UnityEngine.UI;
 public class AnnotationUISaveData : SaveDataContainer
 {
     //public int Id; //unique id for this boundary feature
+    public string ParentCityObject;
     public Vector3 ConnectionPoint;
     public string AnnotationText;
 
@@ -29,19 +31,22 @@ public class AnnotationUI : MonoBehaviour
     private RectTransform myRectTransform;
     [SerializeField]
     private Text numberText;
-    private Annotator annotator;
+    private AnnotationState annotator;
     [SerializeField]
     private InputField inputField;
 
+    private AnnotationUISaveData saveData;
+
     public int Id { get; private set; }
     public bool IsOpen => annotationBody.gameObject.activeInHierarchy;
-
-    private AnnotationUISaveData saveData;
+    public string Text => saveData.AnnotationText;
+    public string ParentCityObject => saveData.ParentCityObject;
+    public Vector3RD ConnectionPointRD => ConvertCoordinates.CoordConvert.UnitytoRD(saveData.ConnectionPoint);
 
     private void Awake()
     {
         myRectTransform = GetComponent<RectTransform>();
-        annotator = GetComponentInParent<Annotator>();
+        annotator = GetComponentInParent<AnnotationState>();
         saveData = new AnnotationUISaveData("undefined");
     }
 
@@ -75,10 +80,11 @@ public class AnnotationUI : MonoBehaviour
         annotator.RemoveAnnotation(Id);
     }
 
-    public void Initialize(int id, Vector3 connectionPoint)
+    public void Initialize(int id, string parentCityObjectId, Vector3 connectionPoint)
     {
         SetId(id);
         saveData.ConnectionPoint = connectionPoint;
+        saveData.ParentCityObject = parentCityObjectId;
     }
 
     public void SetId(int id)
