--- 3DAmsterdam/Assets/T3D/Scripts/Annotations/Annotator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac9177a9824264353a5e29117bfbc1fb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
