+++ 3DAmsterdam/Assets/T3D/Scripts/StateMachine/States/AnnotationState.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 672f50b4e93b14c90a8ca7297826ef17
+guid: ac9177a9824264353a5e29117bfbc1fb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
